Time: 160 ms (93.05%), Space: 42.1 MB (77.51%) - LeetHub

+++ 1554-strings-differ-by-one-character/1554-strings-differ-by-one-character.cpp
@@ -11,7 +11,7 @@ public:
         }
         
         int m = dict.size(), n = dict[0].size(); 
-        long mult = 1l; 
+        long long mult = 1; 
         for (int j = n-1; j >= 0; --j) {
             unordered_map<int, vector<int>> seen; 
             for (int i = 0; i < m; ++i) {
